fixed failing tests

+++ src/Amphibian.Patrol.Training.Api/Repositories/TokenRepository.cs
@@ -6,8 +6,9 @@ using System.Threading.Tasks;
 
 using System.Data;
 using Dapper;
-using Dommel;
+using Dapper.Contrib;
 
+using Dapper.Contrib.Extensions;
 using System.Runtime.InteropServices.WindowsRuntime;
 
 namespace Amphibian.Patrol.Training.Api.Repositories
@@ -23,7 +24,7 @@ namespace Amphibian.Patrol.Training.Api.Repositories
 
         public async Task InsertToken(Token token)
         {
-            token.Id = (int)await _connection.InsertAsync(token).ConfigureAwait(false);
+            token.Id = await _connection.InsertAsync(token).ConfigureAwait(false);
         }
 
         public async Task UpdateToken(Token token)
